Add support for specifying hostname alongside port via AppConfiguration
I am building a web server that supports a local browser-based app. I
do not want that web server to be accessible to other devices on the
network (e.g. other people at a coffeeshop). So, as a start, I’d like
to specify the hostname that the wasabi HTTPServer binds to be
127.0.0.1 instead of the default wildcard (0.0.0.0).

I am leaving the default hostname as null, which uses the previous
behavior. I believe this is consistent with NodeJS’s defaults, so I
don’t have motivation to change the defaults (athough, it would be more
conservative to not assume the user wants their in-development web
server to be accessible to other devices on the network).

I did however change the AppServer used during tests to use 127.0.0.1
instead of 0.0.0.0. Since tests are run locally, this change didn’t
require any other changes. Hypothetically, this would prevent tests
randomly being influenced by some unexpected traffic on your network
(again e.g., someone port scanning at a coffee shop).

Unfortunately, I could not think of a way to test the negative - that
servers are inaccessible from outside of localhost. I did test this
manually in my real-world app, though.